+++ src/demo/utils/demos/customStyles.vue
@@ -1,9 +1,11 @@
 <template>
   <!-- You can open the search component with Ctrl + k or Command + k -->
   <tv-search
-    placeholder="Type 'How'..."
+    placeholder="Type 'Blog'..."
     titleButton="Search"
     :results="results"
+    :customStyles="customStyles"
+    @search="console.log('Searching for:', $event)"
   />
 </template>
 
@@ -11,34 +13,41 @@
 import { ref } from 'vue';
 import { TvSearch } from '@todovue/tv-search';
 
+const customStyles = ref({
+  bgBody: "#0A4539",
+  bgInput: "#284780",
+  bgButton: "#80286E",
+  colorButton: "#D5B7B7",
+});
+
 const results = ref([
   {
     id: 1,
-    title: 'How to use Vue 3',
+    title: 'Blog - How to use Vue 3',
     description: 'Vue 3 is the latest version of Vue.js',
     url: 'https://todovue.com/blog/how-to-use-vue-3',
   },
   {
     id: 2,
-    title: 'How to use Vite',
+    title: 'Blog - How to use Vite',
     description: 'Vite is a build tool for modern web development',
     url: 'https://todovue.com/blog/how-to-use-vite',
   },
   {
     id: 3,
-    title: 'How to use Pinia',
+    title: 'Blog - How to use Pinia',
     description: 'Pinia is a modern store for Vue 3',
     url: 'https://todovue.com/blog/how-to-use-pinia',
   },
   {
     id: 4,
-    title: 'How to use Windi CSS',
+    title: 'Blog - How to use Windi CSS',
     description: 'Windi CSS is a utility-first CSS framework',
     url: 'https://todovue.com/blog/how-to-use-windi-css',
   },
   {
     id: 5,
-    title: 'How to use Vitesse',
+    title: 'Blog - How to use Vitesse',
     description: 'Vitesse is a Vue 3 starter template',
     url: 'https://todovue.com/blog/how-to-use-vitesse',
   }