+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to `@todovue/tv-search` will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/).
 
----
+## [1.0.3] - 2025-11-13
+### 🐛 Fixed
+- The emit event is fixed to send the selected object and not the input.
+
+### 🛠️ Changed
+- Corrected demo application routes.
+
 ## [1.0.2] - 2025-10-17
 ### 🛠️ Changed
 - The library build now uses `src/entry.ts` (exports both the component and the plugin) instead of directly exporting the `.vue` file.
@@ -17,7 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Documentation for usage in SSR and Nuxt 3 applications.
 
 ## [1.0.1] - 2025-05-07
-### 🐛 Bug Fixes
+### 🐛 Fixed
 - Rename `global.scss` to `_global.scss`
 - Update imports in component
 
@@ -34,6 +40,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Easy integration into any Vue 3 application.
 - Published to [npm](https://www.npmjs.com/package/@todovue/tv-search) under the `@todovue` namespace.
 
+[1.0.3]: https://github.com/TODOvue/tv-button/pull/5/files
 [1.0.2]: https://github.com/TODOvue/tv-button/pull/4/files
 [1.0.1]: https://github.com/TODOvue/tv-button/pull/2/files
 [1.0.0]: https://github.com/TODOvue/tv-button/pull/2/files